--- packages/doctor_pet/lib/views/login/login.dart
@@ -1,173 +1,0 @@
-// main.dart
-import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart';
-
-class loginView extends StatelessWidget {
-  const loginView({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        primarySwatch: Colors.blue,
-      ),
-      home: const SignUpScreen(title: 'Đăng ký'),
-    );
-  }
-}
-
-class SignUpScreen extends StatefulWidget {
-  const SignUpScreen({Key? key, required this.title}) : super(key: key);
-
-  final String title;
-
-  @override
-  _SignUpScreenState createState() => _SignUpScreenState();
-}
-
-class _SignUpScreenState extends State<SignUpScreen> {
-  bool isShow = false;
-  final _formKey = GlobalKey<FormState>();
-
-  String? email;
-  String? password;
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-        body: Form(
-      key: _formKey,
-      child: Container(
-        decoration: const BoxDecoration(
-          image: DecorationImage(
-            image: AssetImage(
-              'assets/img/LoginBackground.jpg',
-            ), // Sửa đường dẫn đến ảnh nền
-            fit: BoxFit.cover,
-          ),
-        ),
-        child: Container(
-          width: double.infinity,
-          height: double.infinity,
-          padding: const EdgeInsets.symmetric(horizontal: 20.0),
-          child: Align(
-            alignment: Alignment.centerRight, // Đặt ô đăng ký sang phải
-            child: Container(
-              width: 400,
-              height: 500,
-              padding: const EdgeInsets.all(16),
-              decoration: BoxDecoration(
-                borderRadius: BorderRadius.circular(13),
-                color: Colors.grey[200],
-              ),
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.center,
-                children: [
-                  const Text(
-                    'Login An Account',
-                    style: TextStyle(
-                      fontSize: 20,
-                      fontWeight: FontWeight.bold,
-                    ),
-                  ),
-                  const SizedBox(height: 16.0),
-                  const Text(
-                    'Come in to enjoy our clinic\'s services, you love pets - so do we!!',
-                    style: TextStyle(
-                      fontSize: 12,
-                    ),
-                  ),
-                  const SizedBox(height: 16.0),
-                  TextFormField(
-                    validator: (value) {
-                      if (value!.isEmpty) {
-                        return 'Please enter email';
-                      }
-                      if (!RegExp(
-                              r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
-                          .hasMatch(value)) {
-                        return 'Please enter a valid email';
-                      }
-                      return null;
-                    },
-                    onSaved: (value) => email = value,
-                    decoration: InputDecoration(
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(10),
-                      ),
-                      labelText: 'Email Address',
-                    ),
-                  ),
-                  const SizedBox(height: 16.0),
-                  TextFormField(
-                    validator: (value) {
-                      if (value!.isEmpty) {
-                        return 'Please enter password';
-                      }
-                      if (value.length < 8) {
-                        return 'Password must be at least 8 characters';
-                      }
-                      return null;
-                    },
-                    onSaved: (value) => password = value,
-                    decoration: InputDecoration(
-                      suffixIcon: InkWell(
-                        onTap: () {
-                          setState(() {
-                            isShow = !isShow;
-                          });
-                        },
-                        child: Icon(
-                          isShow ? Icons.visibility_off : Icons.visibility,
-                        ),
-                      ),
-                      border: OutlineInputBorder(
-                        borderRadius: BorderRadius.circular(10),
-                      ),
-                      labelText: 'Password',
-                    ),
-                    obscureText: !isShow,
-                  ),
-                  const SizedBox(height: 16.0),
-                  ElevatedButton(
-                    onPressed: () {
-                      if (_formKey.currentState!.validate()) {
-                        _formKey.currentState!.save();
-                        // Process data
-                      }
-                    },
-                    child: const Text('Login'),
-                  ),
-                  const SizedBox(height: 16.0),
-                  Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      TextButton(
-                        onPressed: () {
-                          // Handle forgot password button press
-                        },
-                        child: const Text('Forgot Password'),
-                      ),
-                      Row(
-                        children: [
-                          const Text('New to us?'),
-                          TextButton(
-                            onPressed: () {
-                              // Handle sign-up button press
-                            },
-                            child: const Text('Sign Up'),
-                          ),
-                        ],
-                      ),
-                    ],
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    ));
-  }
-}
